requirements.txt
# API
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
email-validator==2.1.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Elasticsearch
elasticsearch[async]==8.13.0

# Redis / Celery
redis==5.2.1
celery[redis]==5.4.0

# IMAP
aioimaplib==1.1.0

# Email parsing
mail-parser==3.15.0

# Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==44.0.0

# AI providers
anthropic==0.40.0
openai==1.58.1
httpx==0.28.1

# Embeddings (local)
sentence-transformers==3.3.1

# Telegram
python-telegram-bot==21.9

# Config
pydantic-settings==2.7.0
python-dotenv==1.0.1

# Utils
markdown-it-py==3.0.0
pyyaml==6.0.2
